+++ src/Makefile.am
@@ -1,19 +1,24 @@
 SUBDIRS = . test
 
-bin_PROGRAMS = ior mdtest
+bin_PROGRAMS = ior mdtest md-workbench
 if USE_CAPS
-bin_PROGRAMS += IOR MDTEST
+bin_PROGRAMS += IOR MDTEST MD-WORKBENCH
 endif
 
-noinst_HEADERS = ior.h utilities.h parse_options.h aiori.h iordef.h ior-internal.h option.h mdtest.h aiori-debug.h aiori-POSIX.h
+noinst_HEADERS = ior.h utilities.h parse_options.h aiori.h iordef.h ior-internal.h option.h mdtest.h aiori-debug.h aiori-POSIX.h md-workbench.h
 
 lib_LIBRARIES = libaiori.a
-libaiori_a_SOURCES = ior.c mdtest.c utilities.c parse_options.c ior-output.c option.c
+libaiori_a_SOURCES = ior.c mdtest.c utilities.c parse_options.c ior-output.c option.c md-workbench.c
 
 extraSOURCES = aiori.c aiori-DUMMY.c
 extraLDADD =
 extraLDFLAGS =
 extraCPPFLAGS =
+
+md_workbench_SOURCES = md-workbench.c md-workbench-main.c
+md_workbench_LDFLAGS =
+md_workbench_LDADD = libaiori.a
+md_workbench_CPPFLAGS =
 
 ior_SOURCES = ior-main.c
 ior_LDFLAGS =
